backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
pdfplumber>=0.10.2
openai==0.28.1
python-dotenv>=1.0.0
werkzeug>=3.0.0
httpx>=0.24.1
requests==2.31.0
beautifulsoup4==4.12.2
